requirements.txt
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
git+https://github.com/openai/CLIP.git
easyocr==1.4.2
geotext==0.4.0
huggingface_hub==0.7.0
langdetect==1.0.9
lingua-language-detector==1.0.1
numpy==1.23.0
Pillow==9.3.0
pycountry==22.3.5
PySimpleGUI==4.60.4
requests==2.27.1
scikit_image==0.19.2
scikit_learn==1.1.1
Shapely==1.8.2 
torch==1.12.0 --extra-index-url https://download.pytorch.org/whl/cu113
tqdm==4.64.0
transformers==4.19.4
yolov5==6.1.6
requests==2.27.1

